Updated to using decimal instead of double

+++ CommonNetCoreFuncs/Conversion/UnitConversion.cs
@@ -5,15 +5,15 @@
 /// </summary>
 public static class UnitConversion
 {
-    public const double KgToLbsConst = 2.20462;
-    public const double FtToInConst = 12;
+    public const decimal KgToLbsConst = 2.20462m;
+    public const decimal FtToInConst = 12;
 
     /// <summary>
     /// Convert mass in lbs to kg
     /// </summary>
     /// <param name="massLbs"></param>
-    /// <returns>Double representation of the mass in lbs converted to kg</returns>
-    public static double LbsToKg(this double massLbs)
+    /// <returns>Decimal representation of the mass in lbs converted to kg</returns>
+    public static decimal LbsToKg(this decimal massLbs)
     {
         return massLbs / KgToLbsConst;
     }
@@ -22,12 +22,12 @@ public static class UnitConversion
     /// Convert mass in lbs to kg
     /// </summary>
     /// <param name="massLbs"></param>
-    /// <returns>Double representation of the mass in lbs converted to kg</returns>
-    public static double LbsToKg(this double? massLbs)
+    /// <returns>Decimal representation of the mass in lbs converted to kg</returns>
+    public static decimal LbsToKg(this decimal? massLbs)
     {
         if (massLbs != null)
         {
-            return Convert.ToDouble(massLbs) / KgToLbsConst;
+            return Convert.ToDecimal(massLbs) / KgToLbsConst;
         }
         else
         {
@@ -39,8 +39,8 @@ public static class UnitConversion
     /// Convert mass in kg to lbs
     /// </summary>
     /// <param name="massKg"></param>
-    /// <returns>Double representation of the mass in kg converted to lbs</returns>
-    public static double KgToLbs(this double massKg)
+    /// <returns>Decimal representation of the mass in kg converted to lbs</returns>
+    public static decimal KgToLbs(this decimal massKg)
     {
         return massKg * KgToLbsConst;
     }
@@ -49,12 +49,12 @@ public static class UnitConversion
     /// Convert mass in kg to lbs
     /// </summary>
     /// <param name="massKg"></param>
-    /// <returns>Double representation of the mass in kg converted to lbs</returns>
-    public static double KgToLbs(this double? massKg)
+    /// <returns>Decimal representation of the mass in kg converted to lbs</returns>
+    public static decimal KgToLbs(this decimal? massKg)
     {
         if (massKg != null)
         {
-            return Convert.ToDouble(massKg) * KgToLbsConst;
+            return Convert.ToDecimal(massKg) * KgToLbsConst;
         }
         else
         {
@@ -66,8 +66,8 @@ public static class UnitConversion
     /// Convert length in inches to feet
     /// </summary>
     /// <param name="lenIns"></param>
-    /// <returns>Double representation of the length in inches converted to feet</returns>
-    public static double InsToFt(this double lenIns)
+    /// <returns>Decimal representation of the length in inches converted to feet</returns>
+    public static decimal InsToFt(this decimal lenIns)
     {
         return lenIns / FtToInConst;
     }
@@ -76,12 +76,12 @@ public static class UnitConversion
     /// Convert length in inches to feet
     /// </summary>
     /// <param name="lenIns"></param>
-    /// <returns>Double representation of the length in inches converted to feet</returns>
-    public static double InsToFt(this double? lenIns)
+    /// <returns>Decimal representation of the length in inches converted to feet</returns>
+    public static decimal InsToFt(this decimal? lenIns)
     {
         if (lenIns != null)
         {
-            return Convert.ToDouble(lenIns) / FtToInConst;
+            return Convert.ToDecimal(lenIns) / FtToInConst;
         }
         else
         {
@@ -93,8 +93,8 @@ public static class UnitConversion
     /// Convert length in feet to inches
     /// </summary>
     /// <param name="lenIns"></param>
-    /// <returns>Double representation of the length in feet converted to inches</returns>
-    public static double FtToIns(this double lenIns)
+    /// <returns>Decimal representation of the length in feet converted to inches</returns>
+    public static decimal FtToIns(this decimal lenIns)
     {
         return lenIns * FtToInConst;
     }
@@ -103,12 +103,12 @@ public static class UnitConversion
     /// Convert length in feet to inches
     /// </summary>
     /// <param name="lenIns"></param>
-    /// <returns>Double representation of the length in feet converted to inches</returns>
-    public static double FtToIns(this double? lenIns)
+    /// <returns>Decimal representation of the length in feet converted to inches</returns>
+    public static decimal FtToIns(this decimal? lenIns)
     {
         if (lenIns != null)
         {
-            return Convert.ToDouble(lenIns) * FtToInConst;
+            return Convert.ToDecimal(lenIns) * FtToInConst;
         }
         else
         {
@@ -120,29 +120,29 @@ public static class UnitConversion
     /// Convert bits to Mb
     /// </summary>
     /// <param name="bits"></param>
-    /// <returns>Double representation of the number of bits in Mb</returns>
-    public static double BitsToMb(this int bits)
+    /// <returns>Decimal representation of the number of bits in Mb</returns>
+    public static decimal BitsToMb(this int bits)
     {
-        return Math.Round(bits / 1048576.0, 1, MidpointRounding.AwayFromZero);
+        return Math.Round(bits / 1048576m, 1, MidpointRounding.AwayFromZero);
     }
 
     /// <summary>
     /// Convert bits to Mb
     /// </summary>
     /// <param name="bits"></param>
-    /// <returns>Double representation of the number of bits in Mb</returns>
-    public static double BitsToMb(this long bits)
+    /// <returns>Decimal representation of the number of bits in Mb</returns>
+    public static decimal BitsToMb(this long bits)
     {
-        return Math.Round(bits / 1048576.0, 1, MidpointRounding.AwayFromZero);
+        return Math.Round(bits / 1048576m, 1, MidpointRounding.AwayFromZero);
     }
 
     /// <summary>
     /// Convert Mb to Gb
     /// </summary>
     /// <param name="mb"></param>
-    /// <returns>Double representation of the number of Mb in Gb</returns>
-    public static double MbToGb(this double mb)
+    /// <returns>Decimal representation of the number of Mb in Gb</returns>
+    public static decimal MbToGb(this decimal mb)
     {
-        return Math.Round(mb / 1024.0, 1, MidpointRounding.AwayFromZero);
+        return Math.Round(mb / 1024m, 1, MidpointRounding.AwayFromZero);
     }
 }
